backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
pymongo==4.6.0
flask-pymongo==2.3.0
certifi==2025.7.14
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
gevent==23.9.1
Werkzeug==2.3.7
